+++ src/main.rs
@@ -1,9 +1,10 @@
 mod cli;
 use clap::Parser;
 use cli::Cli;
+use regex::Regex;
 
 mod parse;
-use parse::{parse_search_term, process_json_input};
+use parse::{parse_search_path, process_json_input};
 
 use std::fs;
 use std::io::{self, BufReader, Read};
@@ -20,7 +21,7 @@ fn process_file(
     file_path: &str,
     field_path_parts: &[&str],
     field_name: &str,
-    expected_values: &[&str],
+    search_regex: &Regex,
     single: bool,
     hide_value: bool,
 ) -> Vec<String> {
@@ -30,7 +31,7 @@ fn process_file(
                 file_content,
                 field_path_parts,
                 field_name,
-                expected_values,
+                search_regex,
                 single,
                 hide_value,
             );
@@ -47,7 +48,7 @@ fn handle_file_input(
     json_files: &Vec<String>,
     field_path_parts: &[&str],
     field_name: &str,
-    expected_values: &[&str],
+    search_regex: &Regex,
     single: bool,
     path_output: bool,
     hide_value: bool,
@@ -57,7 +58,7 @@ fn handle_file_input(
             file_path,
             field_path_parts,
             field_name,
-            expected_values,
+            search_regex,
             single,
             hide_value,
         );
@@ -78,7 +79,7 @@ fn handle_string_or_stdin_input(
     json_string: &Option<String>,
     field_path_parts: &[&str],
     field_name: &str,
-    expected_values: &[&str],
+    search_regex: &Regex,
     single: bool,
     hide_value: bool,
 ) {
@@ -97,7 +98,7 @@ fn handle_string_or_stdin_input(
         json_input_raw,
         field_path_parts,
         field_name,
-        expected_values,
+        search_regex,
         single,
         hide_value,
     );
@@ -108,40 +109,46 @@ fn handle_string_or_stdin_input(
 
 fn main() {
     let args = Cli::parse();
-    let search_term_raw = args.search_term;
+    let search_path_raw = args.search_path;
+    let search_term = args.search_term;
     let json_files = args.json_files;
     let json_string = args.json_string;
     let single = args.single;
     let path_output = args.path_output;
     let field_path_separator = args.field_path_separator;
-    let value_separator = args.value_separator;
     let hide_value = args.hide_value;
 
-    match parse_search_term(&search_term_raw, &field_path_separator, &value_separator) {
-        Ok((field_path_parts, field_name, expected_values)) => {
-            if !json_files.is_empty() {
-                handle_file_input(
-                    &json_files,
-                    &field_path_parts,
-                    field_name,
-                    &expected_values,
-                    single,
-                    path_output,
-                    hide_value,
-                );
-            } else {
-                handle_string_or_stdin_input(
-                    &json_string,
-                    &field_path_parts,
-                    field_name,
-                    &expected_values,
-                    single,
-                    hide_value,
-                );
+    match Regex::new(&search_term) {
+        Ok(search_regex) => match parse_search_path(&search_path_raw, &field_path_separator) {
+            Ok((field_path_parts, field_name)) => {
+                if !json_files.is_empty() {
+                    handle_file_input(
+                        &json_files,
+                        &field_path_parts,
+                        field_name,
+                        &search_regex,
+                        single,
+                        path_output,
+                        hide_value,
+                    );
+                } else {
+                    handle_string_or_stdin_input(
+                        &json_string,
+                        &field_path_parts,
+                        field_name,
+                        &search_regex,
+                        single,
+                        hide_value,
+                    );
+                }
             }
-        }
+            Err(e) => {
+                eprintln!("Error parsing search path: {}", e);
+                std::process::exit(1);
+            }
+        },
         Err(e) => {
-            eprintln!("Error parsing search term: {}", e);
+            eprintln!("Error parsing search term as regex: {}", e);
             std::process::exit(1);
         }
     }